% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{rcmivbinary}
\alias{rcmivbinary}
\title{IV approach when X and Y subject to misclassification in binary IV model}
\usage{
rcmivbinary(gbi, x1, x2, y1, y2)
}
\arguments{
\item{G}{A numeric vector of binary instrumental variable.}

\item{X1}{A numeric vector of the first measurement of binary exposure X}

\item{X2}{A numeric vector of the second measurement of binary exposure X}

\item{Y1}{A numeric vector of the first measurement of binary outcome Y}

\item{Y2}{A numeric vector of the second measurement of binary outcome Y}
}
\value{
A vector of causal effect estimate, bootstrap SE and 95% CI
}
\description{
IV approach when X and Y subject to misclassification in binary IV model
}
\author{
Xinhui Liu
}
