% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{rcmivcor}
\alias{rcmivcor}
\title{IV approach when X&Y have dependent measurement error with continuous IV}
\usage{
rcmivcor(g, x1, x2, y1, y2)
}
\arguments{
\item{g}{A numeric vector of instrumental variable.}

\item{x1}{A numeric vector of the first measurement of X}

\item{x2}{A numeric vector of the second measurement of X}

\item{y1}{A numeric vector of the first measurement of Y}

\item{y2}{A numeric vector of the second measurement of Y}
}
\value{
A vector of causal effect estimate, bootstrap SE and 95% CI
}
\description{
IV approach when X&Y have dependent measurement error with continuous IV
}
\author{
Xinhui Liu
}
